Added column values endpoint
